(fix) Make readme example usage use register()
Previously, we were suggesting that helperfunctions be registered in a
class's `__construct()` method. This means the calls to
`helpers->registerFunction()` cannot be fully tested.

Registering the helper functions in the `register()` method instead
makes any helper classes more testable, and is more consistent with the way we
implement other iguana-based classes.